requirements.txt
streamlit==1.38.0
openai==1.47.0
azure-identity==1.17.0